create report utils module
